resolved some cyclic imports problems
