+++ designer/client/src/http/HttpService/index.ts
@@ -1,0 +1,4 @@
+import { HttpService } from "./HttpService";
+
+export * from "./types";
+export default new HttpService();
