+++ designer/client/src/http/HttpService/types.ts
@@ -1,0 +1,155 @@
+import type { AxiosError } from "axios";
+
+import type { CaretPosition2d } from "../../components/graph/node-modal/editors/expression/ExpressionSuggester";
+import type { ProcessStateType, Scenario } from "../../components/Process/types";
+import type { Expression, NodeId, ProcessAdditionalFields, ScenarioGraph, VariableTypes } from "../../types";
+import type { Instant } from "../../types/common";
+
+export type HealthCheckProcessDeploymentType = {
+    status: string;
+    message: null | string;
+    processes: null | Array<string>;
+};
+
+export type HealthCheckResponse = {
+    state: HealthState;
+    error?: string;
+    processes?: string[];
+};
+
+export enum HealthState {
+    ok = "ok",
+    error = "error",
+}
+
+export type FetchProcessQueryParams = Partial<{
+    search: string;
+    categories: string;
+    isFragment: boolean;
+    isArchived: boolean;
+    isDeployed: boolean;
+}>;
+
+export type StatusesType = Record<Scenario["name"], ProcessStateType>;
+
+export interface AppBuildInfo {
+    name: string;
+    gitCommit: string;
+    buildTime: string;
+    version: string;
+    processingType: any;
+}
+
+export type ComponentActionType = {
+    id: string;
+    title: string;
+    icon: string;
+    url?: string;
+};
+
+export type ComponentType = {
+    id: string;
+    name: string;
+    icon: string;
+    componentType: string;
+    componentGroupName: string;
+    categories: string[];
+    actions: ComponentActionType[];
+    usageCount: number;
+    allowedProcessingModes: ProcessingMode[];
+    links: Array<{
+        id: string;
+        title: string;
+        icon: string;
+        url: string;
+    }>;
+    label: string;
+};
+
+export type SourceWithParametersTest = {
+    sourceId: string;
+    parameterExpressions: {
+        [paramName: string]: Expression;
+    };
+};
+
+export type NodesDeploymentData = Record<NodeId, Record<string, string>>;
+
+export type ScenarioGraphSource = {
+    type: ScenarioGraphSourceType;
+    scenarioGraph?: ScenarioGraph;
+    scenarioLabels?: string[];
+    baseScenarioVersionId?: number;
+};
+
+export enum ScenarioGraphSourceType {
+    FROM_GRAPH = "FromGraph",
+}
+
+export type DeployResponse = {
+    deployedScenarioVersionId: number;
+};
+
+export type NodeUsageData = {
+    fragmentNodeId?: string;
+    nodeId: string;
+    type: string;
+};
+
+export type ComponentUsageType = {
+    name: string;
+    nodesUsagesData: NodeUsageData[];
+    isArchived: boolean;
+    isFragment: boolean;
+    processCategory: string;
+    modificationDate: Instant;
+    modifiedBy: string;
+    createdAt: Instant;
+    createdBy: string;
+};
+
+export type NotificationActions = {
+    success(message: string): void;
+    error(message: string, error: string, showErrorText: boolean): void;
+    warn(message: string): void;
+};
+
+export interface PropertiesValidationRequest {
+    name: string;
+    additionalFields: ProcessAdditionalFields;
+}
+
+export interface ExpressionSuggestionRequest {
+    expression: Expression;
+    caretPosition2d: CaretPosition2d;
+    variableTypes: VariableTypes;
+}
+
+export enum ProcessingMode {
+    "streaming" = "Unbounded-Stream",
+    "requestResponse" = "Request-Response",
+    "batch" = "Bounded-Stream",
+}
+
+export interface ScenarioParametersCombination {
+    processingMode: ProcessingMode;
+    category: string;
+    engineSetupName: string;
+}
+
+export interface ScenarioParametersCombinations {
+    combinations: ScenarioParametersCombination[];
+    engineSetupErrors: Record<string, string[]>;
+}
+
+export type ProcessDefinitionDataDictOption = {
+    key: string;
+    label: string;
+};
+
+export type DictOption = {
+    id: string;
+    label: string;
+};
+
+export type ResponseStatus = { status: "success"; data?: any } | { status: "error"; error: AxiosError<string> };
